🐛 fix: skip primary-reveal gate for non-animating rows
Without this short-circuit, non-animating rows (ResultDetailView
turnRow, older non-latest rows in SimulationView / DL Demo) would hide
the chevron during the initial body pass and only reveal it after
`.onAppear` snaps `visibleChars` to target — a one-frame flicker on
row mount.

Also documents that the gate covers chevron presence only; body
visibility (the thought text) remains driven by `showInnerThought`
inside `thoughtSection()`. Folding `showInnerThought` into the gate
would incorrectly nuke chevron presence in collapsed mode.

Tests revised to model the animating vs non-animating split: the
counter-driven branch uses `isLatest: true, charsPerSecond: 60` (or
`streamingPrimary` for the streaming-buffer tests); a new test pins
the non-animating short-circuit; the nil-primary edge moves under an
explicit animating-row label. Tests split into a sibling extension
file per `.claude/rules/testing.md` (added section crossed the 400-
line `file_length` cap).
