+++ Pastura/Pastura/Views/Components/AgentOutputRow.swift
@@ -340,11 +340,12 @@ struct AgentOutputRow: View {
   /// Whether the `▸ THINKING` chevron + optional body should render at
   /// the given reveal counter.
   ///
-  /// Pure over `(resolvedThought presence, primaryText length, visibleChars)`
-  /// so ``AgentOutputRowContractTests`` can exercise both gates (pre-
-  /// reveal hide, post-reveal show, streaming-buffer source) without a
-  /// SwiftUI host. The `@ViewBuilder` body in ``thoughtSection()`` reads
-  /// its own `@State visibleChars`; tests pass the value explicitly.
+  /// Pure over `(shouldAnimate, resolvedThought presence, primaryText
+  /// length, visibleChars)` so ``AgentOutputRowContractTests`` can
+  /// exercise the gate across the animating / non-animating split
+  /// without a SwiftUI host. The `@ViewBuilder` body in
+  /// ``thoughtSection()`` reads its own `@State visibleChars`; tests
+  /// pass the value explicitly.
   ///
   /// **Why a gate at all:** the Sim path naturally suppresses the chevron
   /// until primary completes because `streamingThought` is empty until the
@@ -355,8 +356,23 @@ struct AgentOutputRow: View {
   /// behavior across all four call sites: chevron appears at the moment
   /// the reveal counter reaches the primary buffer length, regardless of
   /// whether thought content is streaming or pre-known.
+  ///
+  /// **Non-animating rows short-circuit to `true`.** Older rows,
+  /// `ResultDetailView` turnRow, and any row with `!shouldAnimate` snap
+  /// `visibleChars` to target inside `.onAppear` — but the first body
+  /// pass runs before `.onAppear` with `visibleChars == 0`, which would
+  /// hide the chevron for one frame and produce a visible flicker on
+  /// row mount. Suppressing the gate in this case keeps non-animating
+  /// rows visually identical to the pre-fix behavior.
+  ///
+  /// **Gate scope is chevron presence only.** Body visibility (the
+  /// thought text itself) is driven independently by `showInnerThought`
+  /// inside ``thoughtSection()`` — `showInnerThought` MUST NOT be folded
+  /// into this gate, or manual chevron taps in collapsed mode would
+  /// nuke chevron presence too.
   func shouldRevealThoughtSection(visibleChars: Int) -> Bool {
     guard let thought = resolvedThought, !thought.isEmpty else { return false }
+    if !shouldAnimate { return true }
     return visibleChars >= (primaryText?.count ?? 0)
   }
