+++ Pastura/PasturaTests/Views/AgentOutputRowContractTests.swift
@@ -265,110 +265,8 @@ struct AgentOutputRowContractTests {
     #expect(expanded.targetLength - collapsed.targetLength == 5)  // = thought
   }
 
-  // MARK: - shouldRevealThoughtSection — primary-reveal completion gate
-  //
-  // Sim's live-streaming path naturally suppresses the `▸ THINKING` chevron
-  // until primary is fully revealed: `streamingThought` is empty until the
-  // LLM emits the `inner_thought` key, so `resolvedThought.isEmpty` short-
-  // circuits `thoughtSection`. Replay paths (DL Demo, Results, past logs)
-  // get `resolvedThought` from `output.innerThought` (a fully-known string)
-  // and used to render the chevron from frame 1 — visibly inconsistent with
-  // Sim. `shouldRevealThoughtSection(visibleChars:)` introduces a counter-
-  // driven gate so all four call sites share the same two-phase reveal:
-  // primary first, then chevron + thought body.
-
-  @Test func shouldRevealThoughtSectionFalseBeforePrimaryFullyRevealed() {
-    let row = AgentOutputRow(
-      agent: "Alice",
-      output: TurnOutput(fields: [
-        "statement": "hello",  // 5 chars
-        "inner_thought": "quiet"
-      ]),
-      phaseType: .speakAll,
-      showAllThoughts: true
-    )
-    #expect(row.shouldRevealThoughtSection(visibleChars: 0) == false)
-    #expect(row.shouldRevealThoughtSection(visibleChars: 4) == false)
-  }
-
-  @Test func shouldRevealThoughtSectionTrueWhenPrimaryFullyRevealed() {
-    let row = AgentOutputRow(
-      agent: "Alice",
-      output: TurnOutput(fields: [
-        "statement": "hello",  // 5 chars
-        "inner_thought": "quiet"
-      ]),
-      phaseType: .speakAll,
-      showAllThoughts: true
-    )
-    #expect(row.shouldRevealThoughtSection(visibleChars: 5) == true)
-    #expect(row.shouldRevealThoughtSection(visibleChars: 10) == true)
-  }
-
-  @Test func shouldRevealThoughtSectionFalseWhenThoughtAbsent() {
-    // Even with primary fully revealed, no thought → no section. Prevents
-    // an empty chevron from appearing on phases / outputs that omit
-    // `inner_thought`.
-    let row = AgentOutputRow(
-      agent: "Alice",
-      output: TurnOutput(fields: ["statement": "hello"]),
-      phaseType: .speakAll,
-      showAllThoughts: true
-    )
-    #expect(row.shouldRevealThoughtSection(visibleChars: 5) == false)
-    #expect(row.shouldRevealThoughtSection(visibleChars: 100) == false)
-  }
-
-  @Test func shouldRevealThoughtSectionUsesStreamingPrimaryLength() {
-    // Streaming row: the gate must compare against the live primary
-    // buffer length, not the parsed-field length, so the chevron pops in
-    // when the streaming buffer is fully revealed — even if the eventual
-    // committed primary is longer or shorter.
-    let row = AgentOutputRow(
-      agent: "Alice",
-      output: TurnOutput(fields: [:]),
-      phaseType: .speakAll,
-      showAllThoughts: true,
-      streamingPrimary: "live",  // 4 chars
-      streamingThought: "thinking"
-    )
-    #expect(row.shouldRevealThoughtSection(visibleChars: 3) == false)
-    #expect(row.shouldRevealThoughtSection(visibleChars: 4) == true)
-  }
-
-  /// Mirrors ``targetLengthPrefersStreamingPrimaryOverParsedPrimary``. Pins
-  /// that `shouldRevealThoughtSection` reads `primaryText` (which short-
-  /// circuits to `streamingPrimary`) rather than `output.primaryText(for:)`
-  /// directly. A future refactor that drops the short-circuit would
-  /// otherwise silently revert the streaming gate to the parsed-field
-  /// length and re-introduce premature chevron pop-in during live
-  /// inference.
-  @Test func shouldRevealThoughtSectionUsesStreamingPrimaryLengthOverParsedPrimary() {
-    let row = AgentOutputRow(
-      agent: "Alice",
-      output: TurnOutput(fields: ["statement": "FINAL TEN"]),  // 9 chars
-      phaseType: .speakAll,
-      showAllThoughts: true,
-      streamingPrimary: "live",  // 4 chars — must win
-      streamingThought: "thinking"
-    )
-    // visibleChars=4 fully reveals the live buffer (4 chars), so the gate
-    // must flip true. If the gate read parsed-primary (9 chars), this
-    // would still be false.
-    #expect(row.shouldRevealThoughtSection(visibleChars: 4) == true)
-  }
-
-  @Test func shouldRevealThoughtSectionTrueWhenPrimaryNil() {
-    // Edge: rows with no primary (rare — fall-through path for outputs
-    // that lack the canonical phase field). `primaryText?.count ?? 0 == 0`
-    // → gate satisfied at visibleChars=0. Chevron then renders from
-    // frame 1, matching pre-fix behavior for primary-less rows.
-    let row = AgentOutputRow(
-      agent: "Alice",
-      output: TurnOutput(fields: ["inner_thought": "musing"]),
-      phaseType: .speakAll,
-      showAllThoughts: true
-    )
-    #expect(row.shouldRevealThoughtSection(visibleChars: 0) == true)
-  }
+  // `shouldRevealThoughtSection(visibleChars:)` tests live in the sibling
+  // file `AgentOutputRowContractTests+ThoughtSectionGate.swift` per
+  // `.claude/rules/testing.md` § "Splitting a Suite Across Files" (this
+  // file crossed the 400-line `file_length` cap when the gate landed).
 }
